requirements.txt
arandr==0.1.11
asgiref==3.6.0
av==10.0.0
Babel==2.10.3
beautifulsoup4==4.11.2
blinker==1.5
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorzero==2.0
cryptography==38.0.4
cupshelpers==1.0
dbus-python==1.3.2
distro==1.8.0
Flask==2.2.2
gpiozero==2.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lgpio==0.2.2.0
libevdev==0.5
lxml==4.9.2
MarkupSafe==2.1.2
meson==1.0.1
more-itertools==8.10.0
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.19
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pygame==2.1.2
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.6.0
python-dotenv==0.21.0
python-prctl==1.8.1
pytz==2022.7.1
pyudev==0.24.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
rpi-lgpio==0.6
RTIMULib==7.2.1
sense-hat==2.6.0
simplejpeg==1.6.6
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
twython==3.8.2
typing_extensions==4.4.0
urllib3==1.26.12
v4l2-python3==0.3.4
webencodings==0.5.1
Werkzeug==2.2.2
zipp==1.0.0
